Admin can create/edit Events - PR revisions and account for generating passes for names shorter than 6 characters
