--- app/assets/stylesheets/events.scss
@@ -1,3 +1,0 @@
-// Place all the styles related to the Events controller here.
-// They will automatically be included in application.css.
-// You can use Sass (SCSS) here: https://sass-lang.com/
