+++ app/javascript/controllers/upload_controller.js
@@ -1,0 +1,10 @@
+import { Controller } from "stimulus"
+
+export default class extends Controller {
+
+    static targets = [ "file" ]
+
+    update(event) {
+        event.target.nextSibling.innerText = event.target.value.replace(/.*(\/|\\)/, '');
+    }
+}
