+++ app/views/devise/mailer/confirmation_instructions.html.erb
@@ -1,9 +1,6 @@
 <p>Welcome <%= @email %> to <%= t 'global.organization_name' %>!</p>
 
-<% if !@resource.first_name.blank? && !@resource.last_name.blank? && @resource.valid_password?(@resource.first_name[0] + @resource.last_name.downcase) %>
-<p>This account has been created for you by an admin.</p>
-<p>Your temporary password is <%= @resource.first_name[0] + @resource.last_name.downcase %>.</p>
-<% end %>
+<p>Your password is <%= @resource.password %>.</p>
 
 <p>You can confirm your account email through the link below:</p>
