+++ app/views/events/index.html.erb
@@ -6,7 +6,7 @@
 <% end %>
 <% @events.each do |event| %>
   <div class="col-sm-6">
-    <div class="card" style="width: 18rem;">
+    <div class="card w-25">
       <!--img class="card-img-top" src="..." alt="Card image"-->
       <div class="card-body">
         <h5 class="card-title"><%= event.title %><span class="badge badge-primary ml-1"><%= event.eventType %></span></h5>
